Update Dockerfile and add usage docs
In order to sync with Greenwave and WaiverDB in terms of
how the container images are built, this will update the
Dockerfile to use Fedora as a base image, add some labels,
and document expected volumes.

Update the README to document how to build and use the
container image and the OpenShift template.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>