Run unit tests in Jenkins instead of waiting for Travis
The waiting for Travis CI groovy code died with an exception I
couldn't solve.

In order to make sure, that unit tests and some linting is run before
building a container image of the application, the unit tests are
going to be executed by the Jenkins job, too.

While Travis is using tox to run these tests, which will pull
dependencies from PyPI, the Jenkins job will make sure that
dependencies are installed from RPMs on a Fedora node.

This way the constellation of app and its dependencies during unit
testing will match the one used for building the container image.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>