Verify namespace in new topic and test case name
Prints warning when an old topic without namespace is encountered but
it's still possible to create new test result with such message.

Prints warning when new topic [1] with namespace is encountered and the
new topic doesn't match the one in test case name. No new test results
are created from such messages.

Note: If "namespace" field in the message contains ".", only the
component before the first "." is expected to be in the topic.

New topic schema is:

    /topic/VirtualTopic.eng.ci.<namespace>.<artifact>.<event>.{queued,running,complete,error}

Example of a new topic with "osci" namespace:

    /topic/VirtualTopic.eng.ci.osci.brew-build.test.complete

JIRA: FACTORY-4845
Signed-off-by: Lukas Holecek <hluk@email.cz>