Avoid propagating frequent exception from consumer
If an exception is propagated from consumer, the same message could be
delivered again later. But if the exception was thrown, for example,
because the message format is incorrect, consecutive run would probably
result in the same exception and this would only delay processing other
messages.

So for common exception, like KeyError or TypeError (but unlike
HTTPError), the consumer will only log the exception and stop it from
propagating.

JIRA: FACTORY-5345
Signed-off-by: Lukas Holecek <hluk@email.cz>